move lang into java package
